Added a rename objects function to the converter
As, for Seaside Hill, I was renaming objects a lot using something I hacked together, so I may as well add it to this.

Also tweaks how names work in general, so objects won't end up with an empty name. Is that important? Not sure, but I don't want it happening.